ExDoc_ (project v0.1.0)

Para crear una nueva pagina debemos crear un modulo y añadirle a este una funcion. Hay que usar las anotaciones @moduledoc """ para que detecte donde comienza nuestro texto. Además de esto con la anotacion @spec podemos definir los tipos de entrada y salida y añadirlo a la página. Podemos añadir los parámetros de la función y por ultimo añadir ejemplos del código.

@moduledoc """

  Aqui podemos añadir el texto que queramos.

  ## Parámetros

    - Variable x e y tienen que ser enteros.

    ## Examples
    iex> ExDoc.suma(2,3)
    5

"""
@spec suma(Integer,Integer) :: Integer
def suma(x,y) do
  IO.puts(x+y)
end

Parámetros

Las variables x e y tienen que ser enteros.

Examples

iex> ExDoc.suma(2,3)
5

Functions

suma(x, y)

Specs

suma(integer(), Integer) :: Integer