Sigilos (project v0.1.0)

Ejemplos

Ejemplo 1: Sigilo que genera lista de caracteres.Es equivalente a encerrar un texto entre comillas simples

iex> ~c/2 + 7 = 9/
'2 + 7 = 9'

Ejemplo 2: Sigilo que genera listas de palabras.

iex> ~w/rojo azul verde/
["rojo", "azul", "verde"]

Ejemplo 3: Sigilos y regex: ~r

iex> Regex.run(~r/[A-Z][a-z]*/, "55_Antonio")
["Antonio"]
iex> Regex.run(~r/[A-Z][a-z]*/, "55_Antonio", return: :index)
[{3, 7}]
iex> Regex.split(~r{-}, "a-b-c")
["a", "b", "c"]
iex> Regex.split(~r{(x)}, "Elixir", include_captures: true)
["Eli", "x", "ir"]
iex> Regex.replace(~r/b/, "abc", "d")
"adc"

Functions

sigil_j(string)

Creamos un sigilo simple que permite separar un String cuyo delimitador es el espacio. La salida es una lista de Strings.

Parámetros

name: String para realizar el split.

Examples

 iex> ~j/hola mundo
 ["hola", "mundo"]