Sigilos (project v0.1.0)

EJEMPLOS

Ejemplo 1: Sigilo que generan lista de caracteres.Es equivalente a encerrar un texto entre comillas simples

iex> ~c/2 + 7 = 9/
'2 + 7 = 9'

Ejemplo 2: Sigilo que genera listas de palabras.

iex> ~w/rojo azul verde/
["rojo", "azul", "verde"]

Ejemplo 3: Sigilos y regex

iex> Regex.run(~r/abc /, "abc ")
["abc "]

iex(3)>

Functions

sigil_j(string)

Creamos un sigilo simple que permite separar un String cuyo delimitador es el espacio. La salida es una lista de Strings.

Parámetros

name: String para realizar el split.

Examples

 iex> ~j/hola mundo
 ["hola", "mundo"]